Moved actions into separate folder
