+++ src/main/java/de/dhbw/modellbahn/application/routing/PathNotPossibleException.java
@@ -1,6 +1,10 @@
 package de.dhbw.modellbahn.application.routing;
 
 public class PathNotPossibleException extends Exception {
+    /**
+     * Path to the destination is not possible.
+     * e.g. because destination is not reachable.
+     */
     public PathNotPossibleException(String message) {
         super(message);
     }
