+++ src/main/java/de/dhbw/modellbahn/application/routing/action/LocSpeedAction.java
@@ -1,4 +1,4 @@
-package de.dhbw.modellbahn.application.routing;
+package de.dhbw.modellbahn.application.routing.action;
 
 import de.dhbw.modellbahn.domain.locomotive.Locomotive;
 import de.dhbw.modellbahn.domain.locomotive.Speed;
