+++ src/main/java/de/dhbw/modellbahn/application/routing/action/RoutingAction.java
@@ -1,4 +1,4 @@
-package de.dhbw.modellbahn.application.routing;
+package de.dhbw.modellbahn.application.routing.action;
 
 public abstract class RoutingAction {
     public abstract void performAction();
