+++ src/test/java/de/dhbw/modellbahn/application/routing/RouteGeneratorTest.java
@@ -3,6 +3,10 @@ package de.dhbw.modellbahn.application.routing;
 import de.dhbw.modellbahn.adapter.moba.communication.ApiService;
 import de.dhbw.modellbahn.adapter.moba.communication.calls.TrackComponentCallsAdapter;
 import de.dhbw.modellbahn.application.port.moba.communication.TrackComponentCalls;
+import de.dhbw.modellbahn.application.routing.action.ChangeSwitchStateAction;
+import de.dhbw.modellbahn.application.routing.action.LocSpeedAction;
+import de.dhbw.modellbahn.application.routing.action.RoutingAction;
+import de.dhbw.modellbahn.application.routing.action.WaitAction;
 import de.dhbw.modellbahn.domain.graph.Distance;
 import de.dhbw.modellbahn.domain.graph.GraphPoint;
 import de.dhbw.modellbahn.domain.graph.NormalSwitch;
